chore(warnings): progress commit - clear C4267 + trivial C4996 sites
Adds explicit static_cast<T>(...) on size_t -> uint16_t/uint32_t/int narrowings (35 sites across controller + view/widgets/compute), replaces vsprintf with vsnprintf in 3 test/demo PrintHeader helpers, and renames POSIX getch to _getch in the demo main. C4267 category at zero; C4996 down to 9 (the 4 strncpy + 5 getenv sites are deferred per WARNINGS.md §4.3 - they need API-contract decisions and a portable wrapper, not auto-sweep). No public API changes; build clean, all 4 test suites pass (173M+ assertions). Updates WARNINGS.md to reflect live state.

+++ src/controller/src/system/rocprofvis_controller_mem_mgmt.cpp
@@ -129,7 +129,7 @@ void MemoryManager::Init(size_t trace_size)
         num_gigabytes >>= 1;
         exponent <<= 1;
     }
-    m_mem_block_size = exponent << 11;
+    m_mem_block_size = static_cast<uint32_t>(exponent << 11);
 
     spdlog::debug("Physical memory = {}!", s_physical_memory_avail);
     spdlog::debug("Memory manager memory allocation block  size = {}!", m_mem_block_size);
@@ -470,7 +470,7 @@ MemoryManager::Allocate(size_t size, rocprofvis_object_type_t type, SegmentTimel
 
     if(current_pool == nullptr)
     {
-        current_pool = new MemoryPool(size, type, m_mem_block_size);
+        current_pool = new MemoryPool(static_cast<uint32_t>(size), type, m_mem_block_size);
         auto& inner_map = m_object_pools[pool_idetifier];
         auto [inner_it, inserted] = inner_map.emplace(current_pool->m_base, current_pool);
         if(inserted)
@@ -695,10 +695,10 @@ BitSet::FindFirstZero() const
             size_t bit_index = __builtin_ctzll(free_bits);
             size_t bit_pos   = i * WORD_SIZE + bit_index;
 #endif
-            if(bit_pos < Size()) return bit_pos;
+            if(bit_pos < Size()) return static_cast<uint32_t>(bit_pos);
         }
     }
-    return Size();
+    return static_cast<uint32_t>(Size());
 }
 
 uint32_t
@@ -712,7 +712,7 @@ BitSet::Count() const {
         total += __builtin_popcountll(word);
 #endif
     }
-    return total;
+    return static_cast<uint32_t>(total);
 }
