+++ src/view/src/widgets/rocprofvis_compute_widget.cpp
@@ -172,12 +172,14 @@ MetricTableBase::GetTableHeight() const
     uint32_t max_rows = 0;
     if(m_max_rows_in_table == 0)
     {
-        max_rows = m_rows.size();
+        max_rows = static_cast<uint32_t>(m_rows.size());
     }
     else
     {
         max_rows =
-            m_rows.size() < m_max_rows_in_table ? m_rows.size() : m_max_rows_in_table;
+            m_rows.size() < m_max_rows_in_table
+                ? static_cast<uint32_t>(m_rows.size())
+                : m_max_rows_in_table;
     }
         
     return style.ScrollbarSize + line_height + (max_rows * line_height);
@@ -439,11 +441,11 @@ MetricTable::Populate(const AvailableMetrics::Table& table,
                 {
                     snprintf(buf, sizeof(buf), "%.2f",
                              metric_value->values.at(value_name));
-                    row.values[value_index + 3].value = buf;
+                    row.values[static_cast<uint32_t>(value_index + 3)].value = buf;
                 }
                 else
                 {
-                    row.values[value_index + 3].value = "";
+                    row.values[static_cast<uint32_t>(value_index + 3)].value = "";
                 }
             }
         }
